chore: Update Rune.ts to exclude unnecessary files from webpack watch
This commit updates the Rune.ts file to exclude unnecessary files from the webpack watch. The ignored patterns in the webpack configuration have been modified to exclude node_modules, .git, and certain file extensions like .js and .d.ts. This change improves the performance of the watch mode by reducing the number of files that webpack needs to monitor for changes.